-added template of reducer -removed unnecessary files
